Move Last.fm-related things to new service category
Last.fm is now a service started when the bot starts, and available to
all commands/hooks via self.bot.lastfm. This is a refactor to support
future commands/hooks.
